+++ timesketch/frontend/src/components/Sketch/CreateViewForm.vue
@@ -33,7 +33,6 @@ limitations under the License.
 import ApiClient from '../../utils/RestApiClient'
 
 export default {
-  name: 'ts-home-sketch-create-form',
   props: [
     'sketchId',
     'currentQueryString',
@@ -62,6 +61,3 @@ export default {
   }
 }
 </script>
-
-<!-- CSS scoped to this component only -->
-<style scoped lang="scss"></style>
