+++ timesketch/frontend/src/components/Sketch/UploadForm.vue
@@ -54,7 +54,6 @@ limitations under the License.
 import ApiClient from '../../utils/RestApiClient'
 
 export default {
-  name: 'ts-sketch-upload-timeline-form',
   data () {
     return {
       form: {
@@ -89,6 +88,3 @@ export default {
   }
 }
 </script>
-
-<!-- CSS scoped to this component only -->
-<style scoped lang="scss"></style>
